# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crew Scheduler API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CrewMember {
  id: number;
  name: string;
  role: string;
  phone?: string | null;
  email?: string | null;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface CreateCrewMemberInput {
  name: string;
  role: string;
  phone?: string | null;
  email?: string | null;
  avatarUrl?: string | null;
}

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  scheduled: "scheduled",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Job {
  id: number;
  title: string;
  location: string;
  scope: string;
  status: JobStatus;
  startDate: string;
  endDate?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type CreateJobInputStatus =
  (typeof CreateJobInputStatus)[keyof typeof CreateJobInputStatus];

export const CreateJobInputStatus = {
  scheduled: "scheduled",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface CreateJobInput {
  title: string;
  location: string;
  scope: string;
  status: CreateJobInputStatus;
  startDate: string;
  endDate?: string | null;
  notes?: string | null;
}

export type JobDetailStatus =
  (typeof JobDetailStatus)[keyof typeof JobDetailStatus];

export const JobDetailStatus = {
  scheduled: "scheduled",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface Task {
  id: number;
  jobId: number;
  title: string;
  description?: string | null;
  status: TaskStatus;
  assignedTo?: CrewMember | null;
  createdAt: string;
}

export interface JobDetail {
  id: number;
  title: string;
  location: string;
  scope: string;
  status: JobDetailStatus;
  startDate: string;
  endDate?: string | null;
  notes?: string | null;
  createdAt: string;
  crew: CrewMember[];
  tasks: Task[];
}

export interface AssignCrewInput {
  crewIds: number[];
}

export type CreateTaskInputStatus =
  (typeof CreateTaskInputStatus)[keyof typeof CreateTaskInputStatus];

export const CreateTaskInputStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface CreateTaskInput {
  title: string;
  description?: string | null;
  status: CreateTaskInputStatus;
  assignedToId?: number | null;
}

export type UpdateTaskInputStatus =
  (typeof UpdateTaskInputStatus)[keyof typeof UpdateTaskInputStatus];

export const UpdateTaskInputStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface UpdateTaskInput {
  title?: string;
  description?: string | null;
  status?: UpdateTaskInputStatus;
  assignedToId?: number | null;
}

export interface Message {
  id: number;
  jobId: number;
  senderName: string;
  content: string;
  photoUrl?: string | null;
  createdAt: string;
}

export interface SendMessageInput {
  senderName: string;
  content: string;
  photoUrl?: string | null;
}

export interface Photo {
  id: number;
  jobId: number;
  url: string;
  caption?: string | null;
  uploadedBy: string;
  createdAt: string;
}

export interface UploadPhotoInput {
  url: string;
  caption?: string | null;
  uploadedBy: string;
}

export interface DashboardSummary {
  totalJobs: number;
  activeJobs: number;
  completedJobs: number;
  totalCrew: number;
  pendingTasks: number;
  completedTasks: number;
}