# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Crew Scheduler API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: crew
    description: Crew member operations
  - name: jobs
    description: Job operations
  - name: tasks
    description: Job task operations
  - name: messages
    description: Group messaging operations
  - name: photos
    description: Photo sharing operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ─── CREW ────────────────────────────────────────────────────────────────────
  /crew:
    get:
      operationId: listCrew
      tags: [crew]
      summary: List all crew members
      responses:
        "200":
          description: List of crew members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrewMember"
    post:
      operationId: createCrewMember
      tags: [crew]
      summary: Create a crew member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCrewMemberInput"
      responses:
        "201":
          description: Created crew member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrewMember"

  /crew/{id}:
    get:
      operationId: getCrewMember
      tags: [crew]
      summary: Get a crew member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Crew member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrewMember"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCrewMember
      tags: [crew]
      summary: Update a crew member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCrewMemberInput"
      responses:
        "200":
          description: Updated crew member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrewMember"
    delete:
      operationId: deleteCrewMember
      tags: [crew]
      summary: Delete a crew member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── JOBS ────────────────────────────────────────────────────────────────────
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job with its assigned crew and tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobInput"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /jobs/{id}/crew:
    get:
      operationId: getJobCrew
      tags: [jobs]
      summary: Get crew assigned to a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assigned crew
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrewMember"
    post:
      operationId: assignCrewToJob
      tags: [jobs]
      summary: Assign crew members to a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignCrewInput"
      responses:
        "200":
          description: Updated crew list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrewMember"

  /jobs/{id}/crew/{crewId}:
    delete:
      operationId: removeCrewFromJob
      tags: [jobs]
      summary: Remove a crew member from a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: crewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ─── TASKS ───────────────────────────────────────────────────────────────────
  /jobs/{id}/tasks:
    get:
      operationId: listJobTasks
      tags: [tasks]
      summary: List tasks for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── MESSAGES ────────────────────────────────────────────────────────────────
  /jobs/{id}/messages:
    get:
      operationId: listJobMessages
      tags: [messages]
      summary: List messages for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PHOTOS ──────────────────────────────────────────────────────────────────
  /jobs/{id}/photos:
    get:
      operationId: listJobPhotos
      tags: [photos]
      summary: List photos for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Photos list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
    post:
      operationId: uploadPhoto
      tags: [photos]
      summary: Upload a photo to a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadPhotoInput"
      responses:
        "201":
          description: Uploaded photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"

  /photos/{id}:
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/upcoming-jobs:
    get:
      operationId: getUpcomingJobs
      tags: [dashboard]
      summary: Get upcoming jobs in the next 7 days
      responses:
        "200":
          description: Upcoming jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    CrewMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - createdAt

    CreateCrewMemberInput:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required:
        - name
        - role

    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: string
        scope:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - location
        - scope
        - status
        - startDate
        - createdAt

    CreateJobInput:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        scope:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - title
        - location
        - scope
        - status
        - startDate

    JobDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: string
        scope:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        crew:
          type: array
          items:
            $ref: "#/components/schemas/CrewMember"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
      required:
        - id
        - title
        - location
        - scope
        - status
        - startDate
        - createdAt
        - crew
        - tasks

    AssignCrewInput:
      type: object
      properties:
        crewIds:
          type: array
          items:
            type: integer
      required:
        - crewIds

    Task:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        assignedTo:
          $ref: "#/components/schemas/CrewMember"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - title
        - status
        - createdAt

    CreateTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        assignedToId:
          type: integer
          nullable: true
      required:
        - title
        - status

    UpdateTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        assignedToId:
          type: integer
          nullable: true

    Message:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        senderName:
          type: string
        content:
          type: string
        photoUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - senderName
        - content
        - createdAt

    SendMessageInput:
      type: object
      properties:
        senderName:
          type: string
        content:
          type: string
        photoUrl:
          type: string
          nullable: true
      required:
        - senderName
        - content

    Photo:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        url:
          type: string
        caption:
          type: string
          nullable: true
        uploadedBy:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - url
        - uploadedBy
        - createdAt

    UploadPhotoInput:
      type: object
      properties:
        url:
          type: string
        caption:
          type: string
          nullable: true
        uploadedBy:
          type: string
      required:
        - url
        - uploadedBy

    DashboardSummary:
      type: object
      properties:
        totalJobs:
          type: integer
        activeJobs:
          type: integer
        completedJobs:
          type: integer
        totalCrew:
          type: integer
        pendingTasks:
          type: integer
        completedTasks:
          type: integer
      required:
        - totalJobs
        - activeJobs
        - completedJobs
        - totalCrew
        - pendingTasks
        - completedTasks