# lib/api-zod/src/generated/types/createCrewMemberInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crew Scheduler API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCrewMemberInput {
  name: string;
  role: string;
  phone?: string | null;
  email?: string | null;
  avatarUrl?: string | null;
}