# lib/api-zod/src/generated/types/crewMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crew Scheduler API
 * OpenAPI spec version: 0.1.0
 */

export interface CrewMember {
  id: number;
  name: string;
  role: string;
  phone?: string | null;
  email?: string | null;
  avatarUrl?: string | null;
  createdAt: Date;
}